+++ kuma-client/src/models/mod.rs
@@ -1,7 +1,5 @@
 #[cfg(feature = "runtime")]
 pub(crate) mod event;
-#[cfg(feature = "runtime")]
-pub(crate) mod response;
 
 pub mod docker_host;
 pub mod maintenance;